refactor(claude): imported-account 관련 파일을 단일 모듈로 병합 (#21)
provider-adapters/src/claude/
  build-imported-account.js
  resolve-imported-claude-accounts.js
  resolve-imported-claude-snapshot.js
  select-claude-accounts-source.js
  → claude-imported-account.js (4개 → 1개)

agent/src/auth/
  create-claude-imported-account-payload.js
  prepare-claude-imported-account.js
  import-claude-account-into-store.js
  → claude-imported-account.js (3개 → 1개)

각 기능별 1-run 파일을 쪼개둔 구조를 응집도 단위로 합친다.
테스트도 동일 도메인을 한 파일로 모아 case 수(217) 유지.
외부 import site는 새 경로로 교체, public API는 동일.

closes #21
