+++ packages/agent/src/auth/index.js
@@ -7,6 +7,4 @@ export * from './port-fallback.js';
 export * from './manual-paste.js';
 export * from './mock-auth-exchange.js';
 export * from './resolve-claude-account.js';
-export * from './create-claude-imported-account-payload.js';
-export * from './prepare-claude-imported-account.js';
-export * from './import-claude-account-into-store.js';
+export * from './claude-imported-account.js';
