+++ packages/agent/src/auth/resolve-claude-account.js
@@ -1,4 +1,4 @@
-import { selectClaudeAccountsSource } from '../../../provider-adapters/src/claude/select-claude-accounts-source.js';
+import { selectClaudeAccountsSource } from '../../../provider-adapters/src/claude/claude-imported-account.js';
 import { resolveAccount } from './account-resolver.js';
 
 /**
