+++ packages/provider-adapters/src/claude/index.js
@@ -6,10 +6,12 @@ export {
 } from './read-claude-credentials.js';
 
 export { mapClaudeCredentials } from './map-claude-credentials.js';
-export { buildImportedClaudeAccount } from './build-imported-account.js';
-export { resolveImportedClaudeAccounts } from './resolve-imported-claude-accounts.js';
-export { selectClaudeAccountsSource } from './select-claude-accounts-source.js';
-export { resolveImportedClaudeSnapshot } from './resolve-imported-claude-snapshot.js';
+export {
+  buildImportedClaudeAccount,
+  resolveImportedClaudeAccounts,
+  selectClaudeAccountsSource,
+  resolveImportedClaudeSnapshot,
+} from './claude-imported-account.js';
 export {
   resolveClaudeUsageSourcePath,
   resolveClaudeUsageSource,
